+++ src/pages/Link/component/FolderList.module.scss
@@ -1,3 +1,5 @@
+@import "../../../components/ui/Modal.module.scss";
+
 .inputGroup {
   position: relative;
   .inputText {
